UPDATE:  numpy.float_ --> float

+++ eelbrain/_ndvar/ndvar.py
@@ -584,7 +584,7 @@ def filter_data(
         data = ndvar.x.swapaxes(axis, -1)
     sfreq = 1. / ndvar.time.tstep
 
-    if data.dtype != numpy.float_:
+    if data.dtype != float:
         data = data.astype(float)
 
     filter_kwargs.setdefault('copy', True)
